28/03/2024 100%

+++ app/src/main/java/ae/valeto/models/UserInfo.java
@@ -19,7 +19,7 @@ public class UserInfo {
 
     @SerializedName("points")
     @Expose
-    private int points;
+    private Double points;
     @SerializedName("created_at")
     @Expose
     private String createdAt;
@@ -67,11 +67,11 @@ public class UserInfo {
     public void setPicture(String picture) {
         this.picture = picture;
     }
-    public int getPoints () {
+    public Double getPoints () {
         return points;
     }
 
-    public void setPoints(int points) {
+    public void setPoints(Double points) {
         this.points = points;
     }
     public String getCreatedAt() {
